Updating the web server to use a database
Instead of keeping saving products information into a database, in this
commit we installed pg library to connect this web app to a postgresql
database. All the functions and methods that involved product CRUD
operations were refactored to async/await instead callbacks and to
execute queries inside the database.
In order to reproduce the server it is necessary to have a database
named 'node-complete' with the following columns (id serial primary key,
name varchar, price numeric(5,2), description varchar, imageurl varchar)